# Conflicts:
#	app/src/main/java/org/cxct/sportlottery/network/Constants.kt
#	app/src/main/java/org/cxct/sportlottery/network/outright/season/Season.kt
#	app/src/main/java/org/cxct/sportlottery/ui/game/GameViewModel.kt

+++ app/src/main/java/org/cxct/sportlottery/ui/bet/list/BetInfoListParlayAdapter.kt
@@ -71,7 +71,7 @@ class BetInfoListParlayAdapter(private val onTotalQuotaListener: OnTotalQuotaLis
                 sendOutStatus = true
             } else {
 
-                val quota = it.toLong()
+                val quota = it.toDouble()
 
                 when {
                     quota > parlayOdd.max -> {
